+++ internal/core/types.go
@@ -28,6 +28,18 @@ type Room struct {
 	UpdatedAt                     time.Time   `json:"updated_at"`                                 // Updation timestamp
 }
 
+// RoomWithoutMembers represents a chatkit room without listing its members.
+type RoomWithoutMembers struct {
+	ID                            string      `json:"id"`                                         // ID assigned to a room
+	CreatedByID                   string      `json:"created_by_id"`                              // User ID that created the room
+	Name                          string      `json:"name"`                                       // Name assigned to the room
+	PushNotificationTitleOverride *string     `json:"push_notification_title_override,omitempty"` // Optionally override Push Notification title
+	Private                       bool        `json:"private"`                                    // Indicates if room is private or not
+	CustomData                    interface{} `json:"custom_data,omitempty"`                      // Custom data that can be added to rooms
+	CreatedAt                     time.Time   `json:"created_at"`                                 // Creation timestamp
+	UpdatedAt                     time.Time   `json:"updated_at"`                                 // Updation timestamp
+}
+
 type messageIsh interface {
 	isMessageIsh()
 }